WIP: Update k8s-topgun for Helm v3
Behaviour around namespaces changed a lot in Helm 3. Every helm command
is now scoped to a namespace just like kubectl. We now need to specify
the namespace in each helm command.

Tiller is no longer required either, the `init` command was removed
because of this.

TODO: installing a chart to a namespace that doesn't exist currently
fails. There's a proposal to bring this behaviour back:
https://github.com/helm/helm/issues/6794

Signed-off-by: Taylor Silva <tsilva@pivotal.io>